requirements/dev.txt
-r base.txt

django-browser-reload==1.12.1
django-debug-toolbar==4.4.2

# Pre-commit
cfgv==3.4.0
distlib==0.3.8
filelock==3.15.4
identify==2.6.0
nodeenv==1.9.1
pre-commit==3.7.1
PyYAML==6.0.1
virtualenv==20.26.3
